computer_science/kms/examples/pyproject.toml
[project]
name = "kms-examples"
version = "0.1.0"
description = "AWS KMS Python SDK examples"
requires-python = ">=3.12"
dependencies = ["boto3", "cryptography", "python-dotenv"]
